+++ test/ttmlir/Dialect/TTIR/permute/permute_tests_negative.mlir
@@ -7,7 +7,7 @@ module {
   func.func @permute_non_valid_permutation(%arg0: tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16> {
     // CHECK: error: 'ttir.permute' op Expected a permutation of {k | 0 <= k < 3} got (0, 1, 0)
     %0 = tensor.empty() : tensor<16x32x64xbf16>
-    %1 = "ttir.permute"(%arg0, %0) <{operand_constraints = [#any_device, #any_device], permutation = array<i32: 0, 1, 0>}> : (tensor<16x32x64xbf16>, tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16>
+    %1 = "ttir.permute"(%arg0, %0) <{operand_constraints = [#any_device, #any_device], permutation = array<i64: 0, 1, 0>}> : (tensor<16x32x64xbf16>, tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16>
     return %1 : tensor<16x32x64xbf16>
   }
 }
@@ -18,7 +18,7 @@ module {
   func.func @permute_non_valid_permutation(%arg0: tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16> {
     // CHECK: error: 'ttir.permute' op Expected a permutation of {k | 0 <= k < 3} got (0, 1)
     %0 = tensor.empty() : tensor<16x32x64xbf16>
-    %1 = "ttir.permute"(%arg0, %0) <{operand_constraints = [#any_device, #any_device], permutation = array<i32: 0, 1>}> : (tensor<16x32x64xbf16>, tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16>
+    %1 = "ttir.permute"(%arg0, %0) <{operand_constraints = [#any_device, #any_device], permutation = array<i64: 0, 1>}> : (tensor<16x32x64xbf16>, tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16>
     return %1 : tensor<16x32x64xbf16>
   }
 }
@@ -30,7 +30,7 @@ module {
   func.func @permute_non_valid_permutation(%arg0: tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16> {
     // CHECK: error: 'ttir.permute' op Expected result shape (16, 64, 32), got (16, 32, 64)
     %0 = tensor.empty() : tensor<16x32x64xbf16>
-    %1 = "ttir.permute"(%arg0, %0) <{operand_constraints = [#any_device, #any_device], permutation = array<i32: 0, 2, 1>}> : (tensor<16x32x64xbf16>, tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16>
+    %1 = "ttir.permute"(%arg0, %0) <{operand_constraints = [#any_device, #any_device], permutation = array<i64: 0, 2, 1>}> : (tensor<16x32x64xbf16>, tensor<16x32x64xbf16>) -> tensor<16x32x64xbf16>
     return %1 : tensor<16x32x64xbf16>
   }
 }
